refactor: move randomVal to a lib
