+++ website/common/script/fns/index.js
@@ -1,7 +1,6 @@
 import handleTwoHanded from './handleTwoHanded';
 import predictableRandom from './predictableRandom';
 import crit from './crit';
-import randomVal from './randomVal';
 import randomDrop from './randomDrop';
 import autoAllocate from './autoAllocate';
 import updateStats from './updateStats';
@@ -11,7 +10,6 @@ module.exports = {
   handleTwoHanded,
   predictableRandom,
   crit,
-  randomVal,
   randomDrop,
   autoAllocate,
   updateStats,
