+++ website/common/script/ops/revive.js
@@ -4,7 +4,7 @@ import _ from 'lodash';
 import {
   NotAuthorized,
 } from '../libs/errors';
-import randomVal from '../fns/randomVal';
+import randomVal from '../libs/randomVal';
 
 // TODO this is only used on the server
 // move out of common?
